requirements.txt
Flask
PyYAML
watchdog
